+++ packages/vuepress-theme-reco/src/client/components/MobileMenus.vue
@@ -1,0 +1,12 @@
+<template>
+  <div class="mobile-menus-container">
+    <NavbarLinks />
+    <div class="appearance"><span>Appearance </span><ToggleDarkModeButton /></div>
+  </div>
+</template>
+
+<script setup lang="ts">
+import NavbarLinks from './NavbarLinks.vue'
+import ToggleDarkModeButton from './ToggleDarkModeButton.vue'
+
+</script>