+++ packages/vuepress-theme-reco/src/client/components/Series.vue
@@ -1,20 +1,16 @@
 <template>
   <aside class="series-container">
-    <NavbarLinks />
-    <template v-for="item in sidebarItems" :key="item.link || item.text">
-      <SeriesItem :item="item" />
-    </template>
+    <SeriesItem v-for="item in sidebarItems" :item="item" :key="item.link || item.text" />
   </aside>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue'
-import { useSidebarItems } from '../composables/'
+import { useSidebarItems } from '../composables'
 import { SeriesItem } from './SeriesItem'
-import NavbarLinks from './NavbarLinks.vue'
 
 export default defineComponent({
-  components: { SeriesItem, NavbarLinks },
+  components: { SeriesItem },
   setup() {
     const sidebarItems = useSidebarItems()
     return {