+++ packages/vuepress-theme-reco/src/client/utils/throttle.ts
@@ -1,0 +1,13 @@
+export const throttle = (func, wait = 500) => {
+  let timer;
+
+  return (...args) => {
+      if (timer) {
+          return
+      }
+      timer = setTimeout(() => {
+          func.apply(func, args);
+          timer = null
+      }, wait)
+  }
+}