+++ packages/@vuepress-reco/plugin-comments/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @vuepress-reco/vuepress-plugin-comments
 
+## 2.0.0-beta.33
+
+### Patch Changes
+
+- fix: optimize the style
+- Updated dependencies
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.33
+
+## 2.0.0-beta.32
+
+### Patch Changes
+
+- chore(theme-reco): fix copy script
+- Updated dependencies
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.32
+
+## 2.0.0-beta.31
+
+### Patch Changes
+
+- refactor: redesign the style of navbar
+- Updated dependencies
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.31
+
 ## 2.0.0-beta.30
 
 ### Patch Changes