+++ packages/vuepress-theme-reco/CHANGELOG.md
@@ -1,5 +1,47 @@
 # vuepress-theme-reco
 
+## 2.0.0-beta.33
+
+### Patch Changes
+
+- fix: optimize the style
+- Updated dependencies
+  - @vuepress-reco/vuepress-plugin-bulletin-popover@2.0.0-beta.33
+  - @vuepress-reco/vuepress-plugin-code-copy@2.0.0-beta.33
+  - @vuepress-reco/vuepress-plugin-comments@2.0.0-beta.33
+  - @vuepress-reco/vuepress-plugin-page@2.0.0-beta.33
+  - @vuepress-reco/vuepress-plugin-vue-preview@2.0.0-beta.33
+  - @vuepress-reco/shared@2.0.0-beta.33
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.33
+
+## 2.0.0-beta.32
+
+### Patch Changes
+
+- chore(theme-reco): fix copy script
+- Updated dependencies
+  - @vuepress-reco/vuepress-plugin-bulletin-popover@2.0.0-beta.32
+  - @vuepress-reco/vuepress-plugin-code-copy@2.0.0-beta.32
+  - @vuepress-reco/vuepress-plugin-comments@2.0.0-beta.32
+  - @vuepress-reco/vuepress-plugin-page@2.0.0-beta.32
+  - @vuepress-reco/vuepress-plugin-vue-preview@2.0.0-beta.32
+  - @vuepress-reco/shared@2.0.0-beta.32
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.32
+
+## 2.0.0-beta.31
+
+### Patch Changes
+
+- refactor: redesign the style of navbar
+- Updated dependencies
+  - @vuepress-reco/vuepress-plugin-bulletin-popover@2.0.0-beta.31
+  - @vuepress-reco/vuepress-plugin-code-copy@2.0.0-beta.31
+  - @vuepress-reco/vuepress-plugin-comments@2.0.0-beta.31
+  - @vuepress-reco/vuepress-plugin-page@2.0.0-beta.31
+  - @vuepress-reco/vuepress-plugin-vue-preview@2.0.0-beta.31
+  - @vuepress-reco/shared@2.0.0-beta.31
+  - @vuepress-reco/tailwindcss-config@2.0.0-beta.31
+
 ## 2.0.0-beta.30
 
 ### Patch Changes